+++ OneBlink.SDK/models/SubmissionDataUrl.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace OneBlink.SDK.Model
+{
+    public class SubmissionDataUrl
+    {
+        public string url { get; set;}
+    }
+}